+++ lib/kampyo/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Kampyo
-  VERSION = '0.2.0'
+  VERSION = '1.0.0'
 end
